requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
phonenumbers==9.0.7
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
